% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snps.R
\name{add_snps}
\alias{add_snps}
\title{Add SNPs to sce}
\usage{
add_snps(sce, mat, thresh = 0.8)
}
\arguments{
\item{sce}{Singnle Cell Experiment object}

\item{mat}{.mxt output from VartTrix}

\item{thresh}{threshold presence of cells}
}
\description{
Add SNPs to sce
}
