% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_genes.R
\name{common_genes}
\alias{common_genes}
\title{Return a list of the top 100 most commonly expressed genes from a SingleCellExperiment object.}
\usage{
common_genes(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object}
}
\value{
names of top 100 most commonly expressed genes.
}
\description{
Return a list of the top 100 most commonly expressed genes from a SingleCellExperiment object.
}
\examples{
sce <- common_genes(sce)

}
