% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commonvariants_1kgenomes_subset}
\alias{commonvariants_1kgenomes_subset}
\title{Sample vcf file}
\format{
An object of class \code{CollapsedVCF} with 2609 rows and 0 columns.
}
\source{
https://cellsnp-lite.readthedocs.io/en/latest/snp_list.html
}
\usage{
data(commonvariants_1kgenomes_subset)
}
\value{
\subsection{\code{commonvariants_1kgenomes_subset}}{

An object of class CollapsedVcf
}
}
\description{
VCF file containing SNPs from a subset of the 1k Genomes common variants
HG38 genome build.
}
\keyword{datasets}
