% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_calls.R
\name{consensus_calls}
\alias{consensus_calls}
\title{Run existing cell hashing algorithms}
\usage{
consensus_calls(sce)
}
\arguments{
\item{sce}{object of class SingleCellExperiment with HTO altExp assay}
}
\value{
Updated SingleCellExperiment object with logical vector indicating training data, data to be classified (complement of training data) and labels.
}
\description{
Run existing cell hashing algorithms
}
\examples{
sce <- consensus_calls(sce)

}
