% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_conf_calls.R
\name{high_conf_calls}
\alias{high_conf_calls}
\title{Run existing cell hashing algorithms}
\usage{
high_conf_calls(sce, assay = "HTO")
}
\arguments{
\item{sce}{Object of class SingleCellExperiment with HTO altExp assay}

\item{assay}{Name of altExp for cell hashing counts to be retrieve from}
}
\value{
Updated SingleCellExperiment object with logical vector indicating training data, data to be classified (complement of training data) and labels.
}
\description{
Run existing cell hashing algorithms
}
\examples{
data(sce)
sce <- high_conf_calls(sce)

}
