% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign.R
\name{reassign}
\alias{reassign}
\title{Reassign cells using knn}
\usage{
reassign(
  sce,
  k = 10,
  seed = 1,
  train_cells = sce$train,
  predict_cells = sce$predict
)
}
\arguments{
\item{sce}{Single Cell Experiment object}

\item{k}{number of neighbours in knn, default to 10}

\item{seed}{set seed for reproducibility, default to 1}

\item{train}{logical vector specifying which cells to use to train classifier}

\item{predict}{logical vector specifying which cells to classify}
}
\description{
Reassign cells using knn
}
