% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_centroid.R
\name{reassign_centroid}
\alias{reassign_centroid}
\title{Title}
\usage{
reassign_centroid(sce, train_cells = sce$train, predict_cells = sce$predict)
}
\arguments{
\item{sce}{}

\item{train_cells}{}

\item{predict_cells}{}
}
\value{
character vector containing reassignments
}
\description{
Title
}
\examples{
l<-reassign_centroi(sce)
}
