% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{SingleCellExperiment object containing multiplexed RNA and HTO data from
six biological smamples}
\format{
An object of class \code{SingleCellExperiment} with 259 rows and 2000 columns.
}
\usage{
data(sce)
}
\value{
\subsection{\code{sce}}{

An object of class SingleCellExperiment
}
}
\description{
Example SingleCellExperiment object containing demultiplexed scRNAseq data
from six donors, used throughout and built upon in demuxSNP workflow.
}
\keyword{datasets}
