% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf}
\alias{vcf}
\title{Sample vcf file}
\format{
\subsection{\code{vcf}}{

A vcf file containing x variants.
}
}
\source{
https://cellsnp-lite.readthedocs.io/en/latest/snp_list.html
}
\usage{
data(vcf)
}
\description{
A subset of the 1k Genomes common variants HG38 vcf file.
}
\keyword{datasets}
