% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign.R
\name{reassign}
\alias{reassign}
\title{Reassign cells using knn}
\usage{
reassign(
  sce,
  k = 10,
  d = 10,
  train_cells = sce$train,
  predict_cells = sce$predict
)
}
\arguments{
\item{sce}{object of class SingleCellExperiment}

\item{k}{number of neighbours used in knn, defaults to 10}

\item{d}{number of doublets per group combination to simulate, defaults to 10}

\item{train_cells}{logical vector specifying which cells to use to train classifier}

\item{predict_cells}{logical vector specifying which cells to classify}
}
\value{
A SingleCellExperiment with updated group assignments called 'knn'
}
\description{
k-nearest neighbour classification of cells.
Training data is intended to be labels of cells confidently called using cell hashing based methods and their corresponding SNPs.
Prediction data can be remaining cells but can also include the training data.
Doublets are simulated by randomly combining 'd' SNP profiles from each grouping combination.
}
\examples{
data(sce, snps)
sce <- high_conf_calls(sce)
sce <- add_snps(sce = sce, mat = snps, thresh = 0.8)
sce <- reassign(sce = sce, k = 10)

}
