% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign.R
\name{reassign}
\alias{reassign}
\title{Reassign cells using knn}
\usage{
reassign(sce, k = 10, train_cells = sce$train, predict_cells = sce$predict)
}
\arguments{
\item{sce}{Single Cell Experiment object}

\item{k}{number of neighbours in knn, default to 10}

\item{train_cells}{logical vector specifying which cells to use to train classifier}

\item{predict_cells}{logical vector specifying which cells to classify}
}
\value{
A SingleCellExperiment with updated group assignments in "knn" metadata
}
\description{
Reassign cells using knn
}
\examples{
sce <- consensus_calls(sce)
sce <- add_snps(sce = sce, mat = snps, thresh = 0.8)
sce <- reassign(sce = sce, k = 10)

}
