% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{SingleCellExperiment object containing multiplexed RNA and HTO data from
six biological smamples}
\format{
\subsection{\code{sce}}{

An object of class SingleCellExperiment
}
}
\usage{
data(sce)
}
\description{
Example SingleCellExperiment object containing demultiplexed scRNAseq data
from six donors, used throughout and built upon in demuxSNP workflow.
}
\keyword{datasets}
