% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{SingleCellExperiment object containing RNA and HTO data}
\format{
\subsection{\code{sce}}{

A data frame with 7,240 rows and 60 columns:
}
}
\source{
n/a for now
}
\usage{
sce
}
\description{
Example SingleCellExperiment object used throughout and built upon in demuxSNP workflow.
}
\examples{
sce <- data(sce)
}
\keyword{datasets}
