% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps}
\alias{snps}
\title{Sample VarTrix output}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2542 rows and 2000 columns.
}
\usage{
data(snps)
}
\value{
\subsection{\code{snps}}{

An object of class matrix
}
}
\description{
A sample output from VarTrix corresponding to the sce SingleCellExperiment
objec for a subset of SNPs located in well observed genes.
}
\keyword{datasets}
