% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps}
\alias{snps}
\title{Sample VarTrix output}
\format{
\subsection{\code{snps}}{

A matrix with 2,000 cells and 2k odd SNPs
}
}
\source{
n/a for now
}
\usage{
snps
}
\description{
A sample output from VarTrix corresponding to the sce SingleCellExperiment object.
}
\keyword{datasets}
