% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vcf.R
\name{subset_vcf}
\alias{subset_vcf}
\title{Title Subset common variants vcf file to only SNPs seen in commonly expressed genes}
\usage{
subset_vcf(vcf_path, sce)
}
\arguments{
\item{vcf_path}{path to common variants vcf file}

\item{sce}{single cell experimen object}
}
\value{
common_snps.vcf vcf file containing subset of SNPs from common variants file seen in commonly expressed genes
}
\description{
Title Subset common variants vcf file to only SNPs seen in commonly expressed genes
}
\examples{
subset_vcf(common_varians.vcf, my_sce)

}
